move non-test code to utils
